requirements.txt
PyYAML
Flask
MarkupSafe
Markups